+++ examples/basic/src/model-provider/llamacpp/llamacpp-chatbot-llama2-text-prompt.ts
@@ -17,7 +17,7 @@ async function main() {
     chat.messages.push({ role: "user", content: userInput });
 
     const model = llamacpp
-      .TextGenerator({
+      .CompletionTextGenerator({
         // use llama-2-7b-chat.GGUF.q4_0.bin
         promptTemplate: llamacpp.prompt.Llama2,
         contextWindowSize: 4096, // Llama 2 context window size