+++ examples/basic/src/model-provider/llamacpp/llamacpp-stream-structure-json-array-grammar-example.ts
@@ -9,7 +9,7 @@ import { z } from "zod";
 async function main() {
   const structureStream = await streamStructure(
     llamacpp
-      .TextGenerator({
+      .CompletionTextGenerator({
         // run openhermes-2.5-mistral-7b.Q4_K_M.gguf in llama.cpp
         promptTemplate: llamacpp.prompt.ChatML,
         maxGenerationTokens: 1024,