+++ examples/basic/src/model-provider/llamacpp/llamacpp-stream-text-vision-with-bakllava-chat-prompt-example.ts
@@ -12,7 +12,7 @@ async function main() {
 
   const textStream = await streamText(
     llamacpp
-      .TextGenerator({
+      .CompletionTextGenerator({
         promptTemplate: llamacpp.prompt.BakLLaVA1,
         maxGenerationTokens: 1024,
         temperature: 0,