+++ examples/basic/src/model-provider/llamacpp/llamacpp-stream-text-with-vicuna-chat-prompt-example.ts
@@ -3,7 +3,7 @@ import { llamacpp, streamText } from "modelfusion";
 async function main() {
   const textStream = await streamText(
     llamacpp
-      .TextGenerator({
+      .CompletionTextGenerator({
         // run https://huggingface.co/TheBloke/vicuna-7B-v1.5-GGUF with llama.cpp
         promptTemplate: llamacpp.prompt.Vicuna,
         contextWindowSize: 2048, // Vicuna v1.5 context window size