+++ examples/basic/src/model-provider/llamacpp/llamacpp-use-tools-openhermes-example.ts
@@ -6,7 +6,7 @@ import { weather } from "../../tool/tools/weather-tool";
 async function main() {
   const { text, toolResults } = await useTools(
     llamacpp
-      .TextGenerator({
+      .CompletionTextGenerator({
         // run https://huggingface.co/TheBloke/OpenHermes-2.5-Mistral-7B-GGUF with llama.cpp
         promptTemplate: llamacpp.prompt.ChatML,
         temperature: 0,